+++ pyhpeimc/plat/termaccess.py
@@ -7,6 +7,7 @@
 import json
 import requests
 import ipaddress
+from pyhpeimc.plat.device import get_dev_details
 
 
 HEADERS = {'Accept': 'application/json', 'Content-Type':
@@ -71,13 +72,14 @@ def get_real_time_locate(ipAddress, auth, url):
                 else:
                     return json.loads(r.text)['realtimeLocation']
             else:
-                return json.loads(r.text)
+                print ("Host not found")
+                return 403
 
     except requests.exceptions.RequestException as e:
             return "Error:\n" + str(e) + " get_real_time_locate: An Error has occured"
 
 
-def get_ip_mac_arp_list(devId, auth,url):
+def get_ip_mac_arp_list(auth,url, devId = None, devip = None):
     """
     function takes devid of specific device and issues a RESTFUL call to get the IP/MAC/ARP list from the target device.
 
@@ -97,15 +99,23 @@ def get_ip_mac_arp_list(devId, auth,url):
 
     >>> auth = IMCAuth("http://", "10.101.0.203", "8080", "admin", "admin")
 
-    >>> ip_mac_list = get_ip_mac_arp_list('10', auth.creds, auth.url)
+    >>> ip_mac_list = get_ip_mac_arp_list( auth.creds, auth.url, devId='10')
+
+    >>> ip_mac_list = get_ip_mac_arp_list( auth.creds, auth.url, devip='10.101.0.221')
 
     >>> assert type(ip_mac_list) is list
 
     >>> assert 'deviceId' in ip_mac_list[0]
 
     """
-    if auth is None or url is None:  # checks to see if the imc credentials are already available
-        set_imc_creds()
+    if devip is not None:
+        dev_details = get_dev_details(devip, auth,url)
+        if type(dev_details) is not str:
+            devId = get_dev_details(devip, auth,url)['id']
+        elif type(dev_details) is str:
+            print ("Device not found")
+            return 403
+
     ip_mac_arp_list_url = "/imcrs/res/access/ipMacArp/" + str(devId)
     f_url = url + ip_mac_arp_list_url
     r = requests.get(f_url, auth=auth, headers=HEADERS)  # creates the URL using the payload variable as the contents
@@ -167,7 +177,7 @@ def get_ip_scope(auth, url, scopeId=None,):
     except requests.exceptions.RequestException as e:
             return "Error:\n" + str(e) + " get_ip_scope: An Error has occured"
 
-def get_ip_scope_detail(scopeId, auth, url ):
+def get_ip_scope_detail( auth, url, scopeId= None, network_address = None ):
     """
     function requires no inputs and returns all IP address scopes currently configured on the HPE IMC server. If the
     optional scopeId parameter is included, this will automatically return only the desired scope id.
@@ -187,15 +197,16 @@ def get_ip_scope_detail(scopeId, auth, url ):
 
     >>> auth = IMCAuth("http://", "10.101.0.203", "8080", "admin", "admin")
 
-    >>> ip_scope_detail = get_ip_scope_detail('45', auth.creds, auth.url)
+    >>> ip_scope_detail = get_ip_scope_detail( auth.creds, auth.url, scopeId = '45')
 
     >>> assert type(ip_scope_detail) is dict
 
     >>> assert 'startIp' in ip_scope_detail
 
     """
+    if network_address != None:
+        scopeId = get_scope_id(network_address, auth, url)
     get_ip_scope_url = "/imcrs/res/access/assignedIpScope/"+str(scopeId)
-
     f_url = url + get_ip_scope_url
     r = requests.get(f_url, auth=auth, headers=HEADERS)  # creates the URL using the payload variable as the contents
     try:
@@ -205,7 +216,7 @@ def get_ip_scope_detail(scopeId, auth, url ):
     except requests.exceptions.RequestException as e:
             return "Error:\n" + str(e) + " get_ip_scope: An Error has occured"
 
-def add_ip_scope(startIp, endIp, name, description, auth, url):
+def add_ip_scope(name, description, auth, url, startIp = None , endIp = None, network_address = None ):
     """
     Function takes input of four strings Start Ip, endIp, name, and description to add new Ip Scope to terminal access
     in the HPE IMC base platform
@@ -249,9 +260,10 @@ def add_ip_scope(startIp, endIp, name, description, auth, url):
 
 
     """
-    if auth is None or url is None:  # checks to see if the imc credentials are already available
-        set_imc_creds()
-
+    if network_address != None:
+        nw_address = ipaddress.IPv4Network(network_address)
+        startIp = nw_address[1]
+        endIp = nw_address[-2]
     add_ip_scope_url = "/imcrs/res/access/assignedIpScope"
     f_url = url + add_ip_scope_url
     payload = ('''{  "startIp": "%s", "endIp": "%s","name": "%s","description": "%s" }'''
@@ -267,7 +279,7 @@ def add_ip_scope(startIp, endIp, name, description, auth, url):
     except requests.exceptions.RequestException as e:
             return "Error:\n" + str(e) + " add_ip_scope: An Error has occured"
 
-def add_child_ip_scope(startIp, endIp, name, description, scopeid, auth, url):
+def add_child_ip_scope(name, description, auth, url,startIp=None, endIp=None, parent_scopeid=None, network_address=None, parent_network_address=None):
     """
     Function takes input of four strings Start Ip, endIp, name, and description to add new Ip Scope to terminal access
     in the HPE IMC base platform
@@ -279,6 +291,8 @@ def add_child_ip_scope(startIp, endIp, name, description, scopeid, auth, url):
     :param name: str Name of the owner of this IP scope  ex. 'admin'
 
     :param description: str description of the Ip scope
+
+    :param parent_scopeid
 
     :param auth: requests auth object #usually auth.creds from auth pyhpeimc.auth.class
 
@@ -296,16 +310,17 @@ def add_child_ip_scope(startIp, endIp, name, description, scopeid, auth, url):
 
     >>> add_child_ip_scope('10.50.0.1', '10.50.0.126', 'cyoung', 'test sub scope', '175', auth.creds, auth.url)
 
-
-
     """
-    if auth is None or url is None:  # checks to see if the imc credentials are already available
-        set_imc_creds()
-
-    add_ip_scope_url = "/imcrs/res/access/assignedIpScope/" + str(scopeid)
+    if parent_network_address != None:
+        parent_scopeid = get_scope_id(parent_network_address, auth, url)
+    if network_address != None:
+        nw_address = ipaddress.IPv4Network(network_address)
+        startIp = nw_address[1]
+        endIp = nw_address[-2]
+    add_ip_scope_url = "/imcrs/res/access/assignedIpScope/" + str(parent_scopeid)
     f_url = url + add_ip_scope_url
     payload = ('''{  "startIp": "%s", "endIp": "%s","name": "%s","description": "%s", "parentId" : "%s"}'''
-               %(str(startIp), str(endIp), str(name), str(description), str(scopeid)))
+               %(str(startIp), str(endIp), str(name), str(description), str(parent_scopeid)))
     r = requests.post(f_url, auth=auth, headers=HEADERS, data=payload) # creates the URL using the payload variable as the contents
     try:
         if r.status_code == 200:
@@ -336,20 +351,23 @@ def delete_ip_scope(network_address, auth, url):
 
     '''
     scope_id = get_scope_id(network_address, auth,url)
+    if (scope_id) == "Scope Doesn't Exist":
+        return scope_id
     delete_ip_address_url = '''/imcrs/res/access/assignedIpScope/'''+str(scope_id)
     f_url = url + delete_ip_address_url
     r = requests.delete(f_url, auth=auth, headers=HEADERS)
     try:
-        return r
         if r.status_code == 204:
             #print("IP Segment Successfully Deleted")
-            return r.status_code
+            return 204
     except requests.exceptions.RequestException as e:
         return "Error:\n" + str(e) + " delete_ip_scope: An Error has occured"
 
+
+
 #Following functions deal with hosts assigned to IP scopes
 
-def add_scope_ip(ipaddress, name, description, scopeid, auth, url):
+def add_scope_ip(ipaddress, name, description, auth, url, scopeid=None, network_address=None):
     """
     Function to add new host IP address allocation to existing scope ID
 
@@ -373,9 +391,13 @@ def add_scope_ip(ipaddress, name, description, scopeid, auth, url):
 
     >>> auth = IMCAuth("http://", "10.101.0.203", "8080", "admin", "admin")
 
-    >>> add_scope_ip('10.50.0.5', 'cyoung', 'New Test Host','175', auth.creds, auth.url)
+    >>> new_host = add_scope_ip('10.50.0.5', 'cyoung', 'New Test Host','175', auth.creds, auth.url)
 
     """
+    if network_address != None:
+        scopeid = get_scope_id(network_address, auth, url)
+        if scopeid == "Scope Doesn't Exist":
+            return scopeid
     new_ip = { "ip": ipaddress,
       "name": name,
       "description": description}
@@ -386,10 +408,10 @@ def add_scope_ip(ipaddress, name, description, scopeid, auth, url):
                       data=payload)  # creates the URL using the payload variable as the contents
     try:
         if r.status_code == 200:
-            #print("IP Scope Successfully Created")
+            #print("IP Host Successfully Created")
             return r.status_code
         elif r.status_code == 409:
-            #print("IP Scope Already Exists")
+            #print("IP Host Already Exists")
             return r.status_code
     except requests.exceptions.RequestException as e:
         return "Error:\n" + str(e) + " add_ip_scope: An Error has occured"
@@ -427,11 +449,9 @@ def remove_scope_ip(hostid, auth, url):
     >>> assert rem_host == 204
 
     """
-    add_scope_ip_url = '/imcrs/res/access/assignedIpScope/ip/'+str(hostid)
-    f_url = url + add_scope_ip_url
-
-    r = requests.delete(f_url, auth=auth, headers=HEADERS,
-                      )
+    delete_scope_ip_url = '/imcrs/res/access/assignedIpScope/ip/'+str(hostid)
+    f_url = url + delete_scope_ip_url
+    r = requests.delete(f_url, auth=auth, headers=HEADERS,                 )
     try:
         if r.status_code == 204:
             #print("Host Successfully Deleted")
@@ -444,7 +464,7 @@ def remove_scope_ip(hostid, auth, url):
 
 
 
-def get_ip_scope_hosts( scopeId, auth, url):
+def get_ip_scope_hosts( auth, url, scopeId=None, network_address=None):
     """
     Function requires input of scope ID and returns list of allocated IP address for the specified scope
 
@@ -479,6 +499,10 @@ def get_ip_scope_hosts( scopeId, auth, url):
     >>> assert 'id' in ip_scope_hosts[0]
 
     """
+    if network_address != None:
+        scopeId = get_scope_id(network_address, auth, url)
+        if scopeId == "Scope Doesn't Exist":
+            return scopeId
     get_ip_scope_url = "/imcrs/res/access/assignedIpScope/ip?size=10000&ipScopeId="+str(scopeId)
     f_url = url + get_ip_scope_url
     r = requests.get(f_url, auth=auth, headers=HEADERS)  # creates the URL using the payload variable as the contents
@@ -496,50 +520,7 @@ def get_ip_scope_hosts( scopeId, auth, url):
     except requests.exceptions.RequestException as e:
             return "Error:\n" + str(e) + " get_ip_scope: An Error has occured"
 
-def add_scope_ip(ipaddress, name, description, scopeid, auth, url):
-    """
-    Function to add new host IP address allocation to existing scope ID
 
-    :param ipaddress:
-
-    :param name: name of the owner of this host
-
-    :param description: Description of the host
-
-    :param auth: requests auth object #usually auth.creds from auth pyhpeimc.auth.class
-
-    :param url: base url of IMC RS interface #usually auth.url from pyhpeimc.auth.authclass
-
-    :return:
-
-    :rtype:
-
-    >>> from pyhpeimc.auth import *
-
-    >>> from pyhpeimc.plat.termaccess import *
-
-    >>> auth = IMCAuth("http://", "10.101.0.203", "8080", "admin", "admin")
-
-    >>> new_host = add_scope_ip('10.50.0.5', 'cyoung', 'New Test Host','175', auth.creds, auth.url)
-
-    """
-    new_ip = { "ip": ipaddress,
-      "name": name,
-      "description": description}
-    add_scope_ip_url = '/imcrs/res/access/assignedIpScope/ip?ipScopeId='+str(scopeid)
-    f_url = url + add_scope_ip_url
-    payload = json.dumps(new_ip)
-    r = requests.post(f_url, auth=auth, headers=HEADERS,
-                      data=payload)  # creates the URL using the payload variable as the contents
-    try:
-        if r.status_code == 200:
-            #print("IP Host Successfully Created")
-            return r.status_code
-        elif r.status_code == 409:
-            #print("IP Host Already Exists")
-            return r.status_code
-    except requests.exceptions.RequestException as e:
-        return "Error:\n" + str(e) + " add_ip_scope: An Error has occured"
 
 def add_host_to_segment(ipaddress, name, description, network_address, auth, url):
     ''' Function to abstract existing add_scope_ip_function. Allows for use of network address rather than forcing human
@@ -574,13 +555,16 @@ def delete_host_from_segment(ipaddress, networkaddress, auth, url):
 
     '''
     host_id = get_host_id(ipaddress, networkaddress, auth, url)
-    remove_scope_ip(host_id, auth.creds, auth.url)
+    delete_host = remove_scope_ip(host_id, auth, url)
+    return delete_host
 
 
 """
 Following Section are Helper functions to help translate human readable IPv4 address to IMC internal keys for working with IP
 scopes and hosts
 """
+
+
 def get_scope_id(network_address, auth, url):
     """
 
@@ -612,15 +596,20 @@ def get_scope_id(network_address, auth, url):
     """
     netaddr = ipaddress.ip_network(network_address)
     scopes = get_ip_scope(auth, url)
-    for i in scopes:
-        if int(i['id']) > 0:
-            if ipaddress.ip_address(i['startIp']) in netaddr and ipaddress.ip_address(i['endIp']) in netaddr:
-                return i['id']
-            if "assignedIpScope" in i:
-                for child in i['assignedIpScope']:
-                    if ipaddress.ip_address(child['startIp']) in netaddr and ipaddress.ip_address(child['endIp']) in netaddr:
+    for parent_scope in scopes:
+        if int(parent_scope['id']) > 0:
+            if "assignedIpScope" in parent_scope:
+                child_scope = parent_scope['assignedIpScope']
+                if type(child_scope) is dict:
+                    child_scope = [child_scope]
+                for scope in child_scope:
+                    if ipaddress.ip_address(scope['startIp']) == netaddr[1] and ipaddress.ip_address(scope['endIp']) == netaddr[-2]:
+                        return scope['id']
+            if ipaddress.ip_address(parent_scope['startIp']) == netaddr[1] and ipaddress.ip_address(parent_scope['endIp']) == netaddr[-2]:
+                return parent_scope['id']
+    return "Scope Doesn't Exist"
 
-                        return child['id']
+
 
 def get_host_id(host_address, network_address, auth, url):
     """
@@ -656,10 +645,16 @@ def get_host_id(host_address, network_address, auth, url):
 
     """
     scope_id = get_scope_id(network_address, auth, url)
-    all_scope_hosts = get_ip_scope_hosts(scope_id, auth, url)
+    if scope_id == "Scope Doesn't Exist":
+        return scope_id
+    all_scope_hosts = get_ip_scope_hosts( auth, url, scopeId= scope_id)
+    if len(all_scope_hosts) == 0:
+        return "Host Doesn't Exist"
     for host in all_scope_hosts:
         if host['ip'] == host_address:
             return host['id']
+        else:
+            return "Host Doesn't Exist"
